The commit hook no longer runs on a temporary directory
The qa-tools are now run on the current directory. This is a compromise
for speed, since speed is the most important factor for adoption and
usage of these tools.

Before, the qa-tools would be run on a temporary directory that
accurately reflected what was going to be committed. This setup is ideal
form a perspective of robustness and representativeness, but was rather
slow.

As a compromise, committers are now reminded when there are unstaged or
untracked changes that could influence qa-tools results. They have to
confirm to continue running the qa-tools and commit.