Reverting to caching composer dir
caching the vendor dir leads to issues when certain
project or ci requirements are dependant on the
composer install call. For instance installation of
binaries in ./bin. To prevent having to cache all
sorts of directories and files, its easier to cache
the composer dir. At the same time, composer install
might require certain extensions to be loaded, thus
the additional custom php config should be loaded
prior to running composer install