Add Composer script for building phars, formatting
Using Composer for this will give us more control over the exact steps to be taken when building.
For instance, we could export a separate hash-file after building if we wanted to.

Also, we could give the user some information about setting phar.readonly setting to "Off" and
what ini-file should be edited.