Hide Composer output unless something fails
The first attempt was based on the assumption that Composer errors are output on standard error, and diagnostic messages (Installing X, Installing Y, etc.) on standard out. It turns out all diagnostic output is written to standard error, unless it is truly an answer to the question (like how `composer licenses` outputs the list of licences on standard out).

The implemented solution hides all Composer output, unless something goes wrong.

The implemented solution became too complex to maintain in the Makefile (it was already quite complex), so this commit moves it to a dedicated build script.