Make system-under-test exit code determination more reliable
@remyhonig was experiencing failing system tests that worked fine on Ubuntu systems; the test harness couldn't reliably determine the system-under-test's exit code. After `strace`ing the hell out of `expect`, I've determined that `expect`'s `wait` command hangs up on the system-under-test (`SIGHUP`), which, in many circumstances, just aborts the PHP script and just returns an exit code of 0. I haven't been able to determine why it did work on Ubuntu, yet not on Mac. It does appear that the `wait` command is not made specifically for making assertions about the spawned process' exit code.

To address the issue, I have replaced the `wait` command with an `expect` command call which expects a certain magic string, which contains the exit code, to appear. While it is a hack, it is very reliable, and someone must be out for breaking things if he/she would use this magic string in QA Tools' output.